// src/generated/api/models/billSplitServiceApplicationFeaturesAuthenticationGoogleLoginGoogleLoginHandlerResponse.ts
/**
 * Generated by orval v8.23.0 🍺
 * Do not edit manually.
 * Bill Split Service API
 * API for bill splitting, Google login, payment tracking and administration.
 * OpenAPI spec version: v1
 */

export interface BillSplitServiceApplicationFeaturesAuthenticationGoogleLoginGoogleLoginHandlerResponse {
  /** @nullable */
  displayName?: string | null;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  avatarUrl?: string | null;
  sessionId?: string;
  /** @nullable */
  accessToken?: string | null;
  /** @nullable */
  refreshToken?: string | null;
  accessTokenExpiresAtUtc?: string;
  refreshTokenExpiresAtUtc?: string;
}